requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
flake8==5.0.4
greenlet==1.1.3
idna==3.4
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.36
urllib3==1.26.12
vk-api==11.9.9
